refactor(api): split api client into modules
